Crypto, crypto everywhere [new] crypto::capi patch DSS CSP (experimental) [new] crypto::keys export DSA, EC keys [new] crypto::kutil import PEM, DER, PKCS#8, CAPI blob & CNG blob when possible, convert it to PKCS#8, or make a PKCS#12 with a PEM or DER certificate [new] dpapi::capi & dpapi::cng export private keys in PVK format for legacy (RSA & DSA) or PKCS#8 for others (like EC) [new] crypto:: & dpapi:: keys & cert functions with more informations

+++ mimikatz/modules/dpapi/packages/kuhl_m_dpapi_creds.c
@@ -19,7 +19,7 @@ NTSTATUS kuhl_m_dpapi_cred(int argc, wchar_t * argv[])
 	{
 		if(kull_m_file_readData(infile, &file, &szFile))
 		{
-			if(szFile >= FIELD_OFFSET(KULL_M_DPAPI_BLOB, dwMasterKeyVersion))
+			if(szFile >= (DWORD) FIELD_OFFSET(KULL_M_DPAPI_BLOB, dwMasterKeyVersion))
 			{
 				isNT5Cred = RtlEqualGuid(file + sizeof(DWORD), &KULL_M_DPAPI_GUID_PROVIDER);
 				kull_m_dpapi_blob_quick_descr(0, isNT5Cred ? file : ((PKUHL_M_DPAPI_ENCRYPTED_CRED) file)->blob);
@@ -167,7 +167,7 @@ void kuhl_m_dpapi_cred_tryEncrypted(LPCWSTR target, LPCBYTE data, DWORD dataLen,
 	DWORD credLen;
 	if(wcsstr(target, L"Microsoft_WinInet_"))
 	{
-		if(dataLen >= FIELD_OFFSET(KULL_M_DPAPI_BLOB, dwMasterKeyVersion))
+		if(dataLen >= (DWORD) FIELD_OFFSET(KULL_M_DPAPI_BLOB, dwMasterKeyVersion))
 		{
 			if(RtlEqualGuid(data + sizeof(DWORD), &KULL_M_DPAPI_GUID_PROVIDER))
 			{
